Corrección de errores en cliente
